Switched from t.decimal to t.integer in database

+++ db/migrate/20091216091138_create_user_entries.rb
@@ -1,7 +1,7 @@
 class CreateUserEntries < ActiveRecord::Migration
   def self.up
     create_table :user_entries do |t|
-      t.decimal :entry_id
+      t.integer :entry_id
       t.references :user
 
       t.timestamps
